holy mother of dtos. moved all my classes to its own files and reorganized a bit
