+++ Controllers/IMochaController.cs
@@ -141,20 +141,3 @@ public class IMochaController : ControllerBase {
         }
     }
 }
-
-public class CandidateInvitation {
-    public int testId { get; set; }
-    public string email { get; set; } = "";
-    public string name { get; set; } = "";
-}
-
-public class IMochaTestInviteResponse {
-    public long testInvitationId { get; set; }
-    public string testUrl { get; set; } = "";
-}
-
-public class IMochaCandidateInvitationBody {
-    public string email { get; set; } = "";
-    public string name { get; set; } = "";
-    public string callbackUrl { get; set;} = "";
-}
