+++ Models/DTOS/CandidateInvitation.cs
@@ -1,0 +1,7 @@
+namespace IPR_BE.Models.DTO;
+
+public class CandidateInvitation {
+    public int testId { get; set; }
+    public string email { get; set; } = "";
+    public string name { get; set; } = "";
+}
