+++ Models/DTOS/IMochaDTOs/IMochaCandidateInvitationBody.cs
@@ -1,0 +1,8 @@
+namespace IPR_BE.Models.DTO;
+
+
+public class IMochaCandidateInvitationBody {
+    public string email { get; set; } = "";
+    public string name { get; set; } = "";
+    public string callbackUrl { get; set;} = "";
+}
