+++ Models/DTOS/IMochaDTOs/IMochaInviteResponse.cs
@@ -1,0 +1,6 @@
+namespace IPR_BE.Models.DTO;
+
+public class IMochaTestInviteResponse {
+    public long testInvitationId { get; set; }
+    public string testUrl { get; set; } = "";
+}
