+++ Models/DTOS/IMochaDTOs/IMochaQuesDifficultyAnalysis.cs
